+++ shopit/urls/shop.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
-from shopit.urls import get_urls
-
-urlpatterns = get_urls('shop')